requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
numpy
google-generativeai
python-dotenv
pytest
streamlit
watchdog
requests
beautifulsoup4
pypdf
